change impl to work with Polarization.dimension as a long value; fix InitDatabase to be able to do both create and upgrade

+++ caom2persistence/src/main/resources/postgresql/caom2.Plane.sql
@@ -72,7 +72,7 @@ create table caom2.Plane
 
 -- polarization
     polarization_states     varchar(34),
-    polarization_dimension  integer,
+    polarization_dimension  bigint,
 
 -- optimisation
     dataReadAccessGroups tsvector default '',
